Rename litex directory to litex_renode
